Refactored xautolock_toggle.sh and more.

+++ bin/scan_dcp7070dw.sh
@@ -8,7 +8,7 @@ prefix="scanned_"
 cd "$out_path"
 files=$(ls ${prefix}[0-9]* 2>/dev/null)
 if [ "$?" -eq 0 ]; then
-	last=$(echo "$files" | sed -e "s|${prefix}||" -e "s|\.png||" | sort -nr | head -1)
+	last=$(echo "$files" | sed -e "s|${prefix}||" -e "s|\.pdf||" | sort -nr | head -1)
 	next=$((last + 1))
 else
 	next=0
@@ -16,5 +16,11 @@ fi
 
 scan_name="${prefix}${next}"
 scanimage --device-name="brother4:net1;dev0" --progress --format=tff > "${scan_name}.tff"
-convert "${scan_name}.tff"  "${scan_name}.png"
-rm "${scan_name}.tff"
+if [ -e "${scan_name}.tff" ]; then
+	convert "${scan_name}.tff"  "${scan_name}.pdf"
+	rm "${scan_name}.tff"
+	echo "${out_path}/${scan_name}.pdf" | xclip -i
+	echo "Scanned image at: ${out_path}/${scan_name}.pdf"
+else
+	echo "Scan failed -- no output." 1>&2
+fi
